backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
pypdfium2==4.30.0
python-docx==1.1.2
python-dateutil==2.9.0.post0
sentence-transformers==3.4.1
scikit-learn==1.6.1
numpy==1.26.4
transformers>=4.41.0,<5.0.0
torch>=2.2.0
Pillow>=10.0.0
accelerate>=0.27.0
supabase==2.13.0
celery==5.4.0
redis==5.2.1
opencv-python-headless==4.10.0.84
ultralytics==8.3.80
spacy>=3.7.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
ollama>=0.2.0
